requirements.txt
alembic~=1.14.1
fastapi[all]~=0.115.8
psycopg2-binary
python-dotenv~=1.0.1
sqlalchemy~=2.0.37
fastapi-users[sqlalchemy]
uvicorn~=0.34.0
asyncpg
redis<5.0.8
gunicorn
pydantic~=2.10.6
starlette~=0.45.3
setuptools
fastapi-cache2[redis]==0.2.2
loguru
apscheduler==3.10.4
pydantic-settings